--- app/src/main/java/woowacourse/shopping/data/CartRepository.kt
@@ -1,18 +1,0 @@
-package woowacourse.shopping.data
-
-import woowacourse.shopping.model.Product
-
-// TODO: Step2 - CartProductDao를 참조하도록 변경
-class CartRepository {
-    private val cartProducts: MutableList<Product> = mutableListOf()
-
-    fun addCartProduct(product: Product) {
-        cartProducts.add(product)
-    }
-
-    fun getAllCartProducts(): List<Product> = cartProducts.toList()
-
-    fun deleteCartProduct(id: Int) {
-        cartProducts.removeAt(id)
-    }
-}
