+++ app/src/main/java/woowacourse/shopping/ui/cart/CartProductViewHolder.kt
@@ -4,7 +4,7 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.databinding.ItemCartProductBinding
-import woowacourse.shopping.model.Product
+import woowacourse.shopping.domain.model.Product
 
 class CartProductViewHolder(
     private val binding: ItemCartProductBinding,
